Update beam_propagation.py
Switched from using "density" to using "concentration" and "composition" to describe our refractive object. "concentration" is a number between zero and one, describing how much of each material is present at each voxel. "composition" is a more abstract quantity between -inf and +inf, with -inf corresponding to a concentration of 0, and +inf corresponding to a concentration of 1. When you're thinking and visualizing, use "concentration", but when you're doing gradient search, use "composition", since it can't go illegal.